@@ -1,5 +1,5 @@
-const sveltePreprocess = require("svelte-preprocess");
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
 
-module.exports = {
-    preprocess: sveltePreprocess(),
+export default {
+    preprocess: vitePreprocess(),
 };
